Fix policy configuration generation (SOC-10883)
The gen_policy.sh script can be used to generate the policy
configuration files needed by Horizon at runtime, by invoking
the oslopolicy-policy-generator tool in the context of services
running on the same node as Horizon.

This commit brings the following improvements to this script and
the way it is used to generate the nova policy file:
  - fix permission handling: the configuration files belonging to
  the target service aren't accessible to the horizon-venv user,
  so the script needs to run the oslopolicy-policy-generator tool
  in the context of a user that has access to those configuration
  files (in this case, the service user)
  - add support for yaml policy file format (will be required for
  Octavia)
  - fix the configuration file location for nova
  - move the extra nova policy rules required by horizon to a
  separate file instead of sed-ing the generated policy file

Most of these improvements are required to make the script reusable
for Octavia.

Change-Id: I1394301dca679360d395f1cfec0f8c36174eee6b